(Refactor) asset images extensions from png to webp
