--- src/assets/images/login/index.js
@@ -1,4 +1,0 @@
-import loginMonster from './loginMonster.png';
-import titleMonster from './titleMonster.png';
-
-export { loginMonster, titleMonster };
