+++ src/assets/images/onboard/index.js
@@ -1,6 +1,6 @@
-import onboard01 from './onboard01.png';
-import onboard02 from './onboard02.png';
-import onboard03 from './onboard03.png';
-import onboard04 from './onboard04.png';
+import onboard01 from './onboard01.webp';
+import onboard02 from './onboard02.webp';
+import onboard03 from './onboard03.webp';
+import onboard04 from './onboard04.webp';
 
 export { onboard01, onboard02, onboard03, onboard04 };
